requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
billiard==4.2.4
blis==1.3.3
build==1.4.0
catalogue==2.0.10
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.13
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
easyocr==1.7.2
faiss-cpu==1.13.2
fastapi==0.128.0
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.0
kombu==5.6.2
kubernetes==35.0.0
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
lazy_loader==0.4
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
neo4j==6.1.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.1
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
pdf2image==1.17.0
pika==1.3.2
pillow==12.1.0
posthog==5.4.0
preshed==3.0.12
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
pybase64==1.4.3
pyclipper==1.4.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyPDF2==3.0.1
PyPika==0.50.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.1
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-alignments==0.9.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.9
SQLAlchemy==2.0.46
srsly==2.5.2
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2026.1.28
tokenizers==0.21.4
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==4.49.0
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
vine==5.1.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.5.0
weasel==0.4.3
websocket-client==1.9.0
websockets==16.0
win32_setctime==1.2.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
